Fix the factory function
